chore: move events messages db to ion_connect folder
