+++ build.yaml
@@ -30,3 +30,4 @@ targets:
             chat_database: lib/app/features/chat/model/database/chat_database.c.dart
             notifications_database: lib/app/features/feed/notifications/data/database/notifications_database.c.dart
             wallets_database: lib/app/features/wallets/data/database/wallets_database.c.dart
+            event_messages_database: lib/app/features/ion_connect/database/event_messages_database.c.dart
