+++ lib/app/features/ion_connect/database/dao/event_messages_dao.c.dart
@@ -2,8 +2,8 @@
 
 import 'package:drift/drift.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/features/feed/bookmarks/data/database/event_messages_database.c.dart';
-import 'package:ion/app/features/feed/bookmarks/data/database/tables/event_messages_table.c.dart';
+import 'package:ion/app/features/ion_connect/database/event_messages_database.c.dart';
+import 'package:ion/app/features/ion_connect/database/tables/event_messages_table.c.dart';
 import 'package:ion/app/features/ion_connect/model/event_reference.c.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
