+++ lib/app/features/ion_connect/repository/event_messages_repository.c.dart
@@ -3,8 +3,8 @@
 import 'package:collection/collection.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/features/feed/bookmarks/data/database/dao/event_messages_dao.c.dart';
-import 'package:ion/app/features/feed/bookmarks/data/database/event_messages_database.c.dart';
+import 'package:ion/app/features/ion_connect/database/dao/event_messages_dao.c.dart';
+import 'package:ion/app/features/ion_connect/database/event_messages_database.c.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart' as ion;
 import 'package:ion/app/features/ion_connect/model/event_reference.c.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
